SQL_TO_VIZ/pyproject.toml
[tool.poetry]
name = "SQL2VIZ"
version = "0.1.0"
description = ""
authors = ["Alen <alen.p@datalabscorp.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.28.0"
fastapi = "^0.110.0"
pydantic = "^2.6.3"
pathlib = "^1.0.1"
asyncio = "^3.4.3"
plotly = "^5.19.0"
pandas = "^2.2.1"
mysql = "^0.0.3"
langchain = "^0.1.11"
sql-metadata = "^2.10.0"
langchain_google_genai = "^0.0.9"
google-generativeai = "^0.3.2"
python-dotenv = "^1.0.1"
mysql-connector-python = "^8.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
